LLaVA-MED/requirements.txt
absl-py==2.1.0
accelerate==1.0.1
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.43.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
datasets==2.18.0
decorator==5.1.1
deepspeed==0.14.0
dill==0.3.8
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.7.0
einops-exts==0.0.4
evaluate==0.4.1
exceptiongroup==1.2.0
executing==2.0.1
fairscale==0.4.0
fastapi==0.110.0
ffmpy==0.3.2
filelock==3.13.1
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.1.3
gitdb==4.0.11
GitPython==3.1.42
gradio==4.21.0
gradio_client==0.12.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.6
importlib_resources==6.3.0
ipdb==0.13.13
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
jsonpickle==3.2.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lightning-utilities==0.11.2
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpi4py_mpich==3.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
munch==4.0.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.9.1
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.19.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
open-clip-torch==2.24.0
opencv-python==4.10.0.82
orjson==3.9.15
packaging==24.0
pandas==2.2.1
parso==0.8.4
peft==0.11.1
pexpect==4.9.0
pillow==10.2.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==5.9.8
ptflops==0.7.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.1
pyarrow-hotfix==0.6
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytorch-lightning==2.2.5
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2024.9.11
requests==2.31.0
responses==0.18.0
rich==13.7.1
rouge-score==0.1.2
rpds-py==0.18.0
ruff==0.3.2
sacred==0.8.5
safetensors==0.4.5
scikit-learn==1.4.1.post1
scipy==1.12.0
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==1.43.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
starlette==0.36.3
svgwrite==1.4.3
sympy==1.12
tensorboardX==2.6.2.2
threadpoolctl==3.3.0
timm==0.6.13
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.1
torchaudio==2.3.1
torchmetrics==1.4.0.post0
torchscale==0.3.0
torchvision==0.17.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.40.2
triton==2.2.0
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.0
wandb==0.16.4
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==11.0.3
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
